🔥 feat : jpql aggregation 함수, string 함수 추가
- concat, substring 와 같은 Multiple Column 을 가지는 FunctionalColumn 추가
- functions 에서는 supplier 만 제공하고, 나머지는 Predictor 처럼 column에서 제공할 수 있도록 수정하는 것을 고려

+++ src/main/java/bback/module/poqh3/impl/FunctionCommand.java
@@ -13,7 +13,7 @@ enum FunctionCommand {
     , YEAR(Integer.class), MONTH(Integer.class), DAY(Integer.class), HOUR(Integer.class), MINUTE(Integer.class), SECOND(Integer.class)
 
     // aggregation
-    , SUM(Long.class), MIN(Long.class), MAX(Long.class), AVG(Long.class), COUNT(Long.class)
+    , SUM(Integer.class), MIN(Integer.class), MAX(Integer.class), AVG(Double.class), COUNT(Long.class)
 
     // string
     , CONCAT(String.class), SUBSTRING(String.class), TRIM(String.class), LOWER(String.class), UPPER(String.class), LENGTH(String.class)
